Delete checkSpaceExist(), Delete FileBasedSchemaManager, Move AdHocSchemaManager to storage/test
