+++ src/graph/UseExecutor.cpp
@@ -26,14 +26,14 @@ void UseExecutor::execute() {
 
     auto spaceName = *sentence_->space();
     // Check from the cache, if space not exists, schemas also not exist
-    auto status = ectx()->schemaManager()->checkSpaceExist(spaceName);
+    auto status = ectx()->schemaManager()->toGraphSpaceID(spaceName);
     if (!status.ok()) {
         DCHECK(onError_);
         onError_(Status::Error("Space not found for `%s'", spaceName.c_str()));
         return;
     }
 
-    auto spaceId = ectx()->schemaManager()->toGraphSpaceID(spaceName);
+    auto spaceId = status.value();
     session->setSpace(*sentence_->space(), spaceId);
     FLOG_INFO("Graph space switched to `%s', space id: %d", sentence_->space()->c_str(), spaceId);
